feat: implement hash
